refactor(dom): extract DomCaretPlacer from DomModel
